% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_participant.R
\name{aggregate_behavioral}
\alias{aggregate_behavioral}
\title{Aggregate behavioral data across participants}
\usage{
aggregate_behavioral(participant_list)
}
\arguments{
\item{participant_list}{List of participant data}
}
\value{
Single tibble with all participants' behavioral trials combined
}
\description{
Aggregate behavioral data across participants
}
