% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_participant.R
\name{aggregate_events}
\alias{aggregate_events}
\title{Aggregate events across participants}
\usage{
aggregate_events(participant_list)
}
\arguments{
\item{participant_list}{List of participant data}
}
\value{
Single tibble with all participants' events combined
}
\description{
Aggregate events across participants
}
