% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_participant.R
\name{aggregate_metadata}
\alias{aggregate_metadata}
\title{Aggregate metadata across participants}
\usage{
aggregate_metadata(participant_list)
}
\arguments{
\item{participant_list}{List of participant data}
}
\value{
Single tibble with all participants' metadata combined
}
\description{
Aggregate metadata across participants
}
