% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_participant.R
\name{aggregate_samples}
\alias{aggregate_samples}
\title{Aggregate samples across participants}
\usage{
aggregate_samples(participant_list)
}
\arguments{
\item{participant_list}{List of participant data (from read_all_participants)}
}
\value{
Single tibble with all participants' samples combined
}
\description{
Aggregate samples across participants
}
