% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_participant.R
\name{aggregate_validation}
\alias{aggregate_validation}
\title{Aggregate validation data across participants}
\usage{
aggregate_validation(participant_list)
}
\arguments{
\item{participant_list}{List of participant data}
}
\value{
Single tibble with all participants' validation metrics combined
}
\description{
Aggregate validation data across participants
}
