% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{apply_calibration_offsets_eager}
\alias{apply_calibration_offsets_eager}
\title{Apply calibration offsets (eager tibble version)}
\usage{
apply_calibration_offsets_eager(samples_df, validation_df, method)
}
\description{
Apply calibration offsets (eager tibble version)
}
\keyword{internal}
