% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{apply_calibration_offsets_lazy}
\alias{apply_calibration_offsets_lazy}
\title{Apply calibration offsets (lazy LazyFrame version)}
\usage{
apply_calibration_offsets_lazy(samples_lf, validation_df, method)
}
\description{
Apply calibration offsets (lazy LazyFrame version)
}
\keyword{internal}
