% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{binarize_on_task_eager}
\alias{binarize_on_task_eager}
\title{Binarize on-task (eager tibble version)}
\usage{
binarize_on_task_eager(samples_df, zone_radii_df)
}
\description{
Binarize on-task (eager tibble version)
}
\keyword{internal}
