% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{binarize_on_task_polars}
\alias{binarize_on_task_polars}
\title{Binarize on-task (lazy polars version)}
\usage{
binarize_on_task_polars(samples_lf, zone_radii_df)
}
\description{
Binarize on-task (lazy polars version)
}
\keyword{internal}
