% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mat_data.R
\name{classify_event_type}
\alias{classify_event_type}
\title{Classify event type from message string}
\usage{
classify_event_type(msgs)
}
\description{
Classify event type from message string
}
\keyword{internal}
