% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{compute_distance_eager}
\alias{compute_distance_eager}
\title{Compute distance (eager tibble version)}
\usage{
compute_distance_eager(samples_df, use_adjusted)
}
\description{
Compute distance (eager tibble version)
}
\keyword{internal}
