% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{compute_distance_polars}
\alias{compute_distance_polars}
\title{Compute distance (lazy polars version)}
\usage{
compute_distance_polars(samples_lf, use_adjusted)
}
\description{
Compute distance (lazy polars version)
}
\keyword{internal}
