% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_behavioral_data}
\alias{extract_behavioral_data}
\title{Extract behavioral data from MAT file}
\usage{
extract_behavioral_data(mat_file, id_prefix = "CSN")
}
\description{
Extract behavioral data from MAT file
}
