% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mat_data.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract session metadata}
\usage{
extract_metadata(subjdata, recordings, participant_id)
}
\description{
Extract session metadata
}
\keyword{internal}
