% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mat_data.R
\name{extract_participant_id}
\alias{extract_participant_id}
\title{Extract participant ID from filename}
\usage{
extract_participant_id(path)
}
\description{
Extract participant ID from filename
}
\keyword{internal}
