% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_recordings_data}
\alias{extract_recordings_data}
\title{Extract recordings data from eyetracker struct}
\usage{
extract_recordings_data(struct_data)
}
\description{
Extract recordings data from eyetracker struct
}
