% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioral_data_models.R
\name{false_alarm_by_response_time_model}
\alias{false_alarm_by_response_time_model}
\title{Model participant false alarms by response time}
\usage{
false_alarm_by_response_time_model(df, random_effects = FALSE)
}
\description{
TODO
}
