% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioral_data_preprocessing.R
\name{get_all_hits_with_reaction_times}
\alias{get_all_hits_with_reaction_times}
\title{Get all participants' hits with reaction times}
\usage{
get_all_hits_with_reaction_times(combined_df)
}
\description{
Creates a dataframe composed of each participant's hit times and
reaction times for those hits, row-by-row with signal times.
}
