% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioral_data_preprocessing.R
\name{get_behavioral_metadata}
\alias{get_behavioral_metadata}
\title{Get behavioral metadata for each participant}
\usage{
get_behavioral_metadata(behavioral_data)
}
\description{
Creates a dataframe composed of each participant's id, signal
probability, clock side, task begin and end.
}
