% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioral_data_preprocessing.R
\name{get_false_alarms_given_responses}
\alias{get_false_alarms_given_responses}
\title{Get false alarms given responses for all participants}
\usage{
get_false_alarms_given_responses(behavioral_data, hits_with_reaction_times)
}
\description{
TODO
}
