% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioral_data_preprocessing.R
\name{get_hit_times}
\alias{get_hit_times}
\title{Get hit times vector from signals and responses per participant}
\usage{
get_hit_times(signal_times, response_times, .interval = 8)
}
\description{
This function uses input of both signal times and response times
from a single participant's extracted behavioral data to determine the hit
indices, if any, and in turn, the hit times.
Get hit timestamp from a vector of signal times and a vector of response times
}
\note{
Use the .interval variable if 8 seconds is not the desired interval
}
