% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioral_data_preprocessing.R
\name{get_hits_given_signals}
\alias{get_hits_given_signals}
\title{Get all participants' hits given signal presence}
\usage{
get_hits_given_signals(behavioral_data)
}
\description{
Creates a dataframe composed of each participant's hit times and
reaction times for those hits, row-by-row with signal times.
}
