% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_id_vector}
\alias{get_id_vector}
\title{Get vector of integer id values}
\usage{
get_id_vector(path_to_dir, min_id = 1, max_id = 1)
}
\arguments{
\item{path_to_dir}{Character, a path to a directory of CSV data at a given stage}

\item{min_id}{Integer, a minimum integer id value. Default is 1.}

\item{max_id}{Integer, a maximum integer id value. Default is 1.}
}
\value{
vector of integer id values
}
\description{
Create and store integer id values in a vector
for participants relative to the existence of that participant's data.
In some cases, participants are not converted from raw MAT data to CSVs
if not considered a complete participant.
}
\examples{
get_id_vector(extracted_eyetracker_data_path, 1, 57)
}
