% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_path_to_id}
\alias{get_path_to_id}
\title{Get file path to a formal participant identifier}
\usage{
get_path_to_id(path_to_dir, id)
}
\arguments{
\item{path_to_dir}{Character, a path to a directory of CSV data at a given stage}

\item{id}{Character, a unique integer identifier}
}
\value{
path to that participant's id
}
\description{
Participants are identified by order of participation, and
data files for each participant should be stored in a dedicated folder,
such as \code{CSN004}. This function constructs the file path for a given stage
of data, i.e., \code{raw} or \code{extracted}.
}
\examples{
get_path_to_id(extracted_eyetracker_data_path, 4)
}
