% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_participant_data}
\alias{load_participant_data}
\title{Load participant data}
\usage{
load_participant_data(path_to_dir, id_vector, filename_csv)
}
\arguments{
\item{path_to_dir}{A path string to a directory of CSV data}

\item{id_vector}{A vector of participant IDs as integers}

\item{filename_csv}{A string name for a file type, including extension.}
}
\value{
list of dataframes
}
\description{
Read a directory of files, or directory of directories of
files specifying a specific commanly-named file, into a dataframe list.
Specify a \code{filename_csv} string if the data type is nested within a participant
folder (e.g., eye-tracking data, such as \code{extracted/eyetracker/CSN004/fsample.csv}).
Otherwise, the method assumes the data is stored as a CSV-per-participant (e.g.,
behavioral data, such as \code{extracted/behavioral/CSN004.csv}).
}
\examples{
load_participant_data(extracted_eyetracker_data_path, id_vector, event_csv)
}
