% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_pretty_df}
\alias{make_pretty_df}
\title{Make dataframe output pretty}
\usage{
make_pretty_df(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
dataframe
}
\description{
When using \code{knitr::kable} for documentation output from RMarkdown,
return a pretty table. Otherwise, return the dataframe as is.
}
\examples{
make_pretty_df(df)
}
