% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mat_data.R
\name{parse_calibration_fields}
\alias{parse_calibration_fields}
\title{Parse calibration validation messages}
\usage{
parse_calibration_fields(events)
}
\description{
Parse calibration validation messages
}
\keyword{internal}
