% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mat_data.R
\name{parse_fevent}
\alias{parse_fevent}
\title{Parse FEVENT structure into events tibble}
\usage{
parse_fevent(fevent, participant_id)
}
\description{
Parse FEVENT structure into events tibble
}
\keyword{internal}
