% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mat_data.R
\name{parse_fsample}
\alias{parse_fsample}
\title{Parse FSAMPLE structure into samples tibble}
\usage{
parse_fsample(fsample, participant_id)
}
\description{
Parse FSAMPLE structure into samples tibble
}
\keyword{internal}
