% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mat_data.R
\name{parse_subjdata}
\alias{parse_subjdata}
\title{Parse subjdata structure into behavioral events}
\usage{
parse_subjdata(subjdata, participant_id)
}
\description{
Parse subjdata structure into behavioral events
}
\keyword{internal}
