% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioral_data_models.R
\name{reaction_time_by_signal_time_model}
\alias{reaction_time_by_signal_time_model}
\title{Model participant reaction time by signal time}
\usage{
reaction_time_by_signal_time_model(df, random_effects = FALSE)
}
\description{
TODO
}
