% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample_samples_eager}
\alias{resample_samples_eager}
\title{Resample gaze samples (eager tibble version)}
\usage{
resample_samples_eager(samples_df, hz_standard, gap_threshold)
}
\description{
Resample gaze samples (eager tibble version)
}
\keyword{internal}
