% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample_samples_polars}
\alias{resample_samples_polars}
\title{Resample gaze samples (lazy polars version)}
\usage{
resample_samples_polars(samples_lf, hz_standard, gap_threshold)
}
\description{
Resample gaze samples (lazy polars version)
}
\keyword{internal}
