% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_csvs_from_mats}
\alias{extract_csvs_from_mats}
\title{Run matlab extractor}
\usage{
extract_csvs_from_mats(mat_file)
}
\arguments{
\item{mat_file}{Character, file name with extension ".m"}
}
\value{
0 for success; 1 for error. see matlbr docs.
}
\description{
If MATLAB is available on system, run script to create CSV
from MAT files
}
\examples{
library(matlabr)
library(here)
extract_csvs_from_mats("edf_mat_to_csv.m")
}
